REF: misplaced set_index tests (#33822)
